+++ apps/ledger-live-desktop/src/newArch/features/AddAccountDrawer/screens/AccountsOnboard/hooks/useOnboardingFlow.ts
@@ -1,7 +1,7 @@
+import { useOnboardingFlow as useOnboardingFlowShared } from "@ledgerhq/live-common/hooks/useAccountOnboarding";
 import { Device } from "@ledgerhq/live-common/hw/actions/types";
 import { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
 import { Account } from "@ledgerhq/types-live";
-import { useOnboardingFlow as useOnboardingFlowShared } from "@ledgerhq/live-common/hooks/useAccountOnboarding";
 import { OnboardingBridge, OnboardingConfig } from "../types";
 
 // Desktop wrapper that adapts Device to deviceId for shared hook
