+++ apps/ledger-live-mobile/src/newArch/features/Accounts/screens/AccountsOnboard/adapters/canton.ts
@@ -1,36 +1,12 @@
-import type { CantonCurrencyBridge } from "@ledgerhq/coin-canton/types";
-import { CantonOnboardProgress, CantonOnboardResult } from "@ledgerhq/coin-canton/types";
+import { isCantonCurrencyBridge, type CantonCurrencyBridge } from "@ledgerhq/coin-canton";
 import { getCurrencyBridge } from "@ledgerhq/live-common/bridge/index";
 import { isTokenCurrency } from "@ledgerhq/live-common/currencies/index";
 import { CryptoCurrency, TokenCurrency } from "@ledgerhq/types-cryptoassets";
-import { map } from "rxjs/operators";
 import { urls } from "~/utils/urls";
 import StepFinish from "../components/StepFinish";
 import StepOnboard from "../components/StepOnboard";
 import { StepFinishFooter, StepOnboardFooter } from "../footers";
-import {
-  OnboardingBridge,
-  OnboardingConfig,
-  OnboardProgress,
-  OnboardResult,
-  StepId,
-  TranslationKeys,
-  UrlConfig,
-} from "../types";
-
-function isCantonCurrencyBridge(bridge: unknown): bridge is CantonCurrencyBridge {
-  if (!bridge || typeof bridge !== "object") {
-    return false;
-  }
-
-  if (!("onboardAccount" in bridge)) {
-    return false;
-  }
-
-  /* eslint-disable @typescript-eslint/no-explicit-any */
-  const bridgeWithProperty = bridge as any;
-  return typeof bridgeWithProperty.onboardAccount === "function";
-}
+import { OnboardingConfig, StepId, TranslationKeys, UrlConfig } from "../types";
 
 export function getCantonBridge(
   currency: CryptoCurrency | TokenCurrency,
@@ -46,29 +22,6 @@ export function getCantonBridge(
     return bridge;
   }
   return null;
-}
-
-export function createCantonOnboardingBridge(cantonBridge: CantonCurrencyBridge): OnboardingBridge {
-  return {
-    onboardAccount: (currency, deviceId, creatableAccount) => {
-      return cantonBridge.onboardAccount(currency, deviceId, creatableAccount).pipe(
-        map(
-          (data: CantonOnboardProgress | CantonOnboardResult): OnboardProgress | OnboardResult => {
-            if ("status" in data) {
-              return { status: data.status };
-            }
-            if ("account" in data && "partyId" in data) {
-              return {
-                account: data.account,
-                metadata: { partyId: data.partyId },
-              };
-            }
-            throw new Error("Invalid Canton onboarding result");
-          },
-        ),
-      );
-    },
-  };
 }
 
 export const cantonTranslationKeys: TranslationKeys = {
