+++ libs/coin-modules/coin-canton/src/index.ts
@@ -1,6 +1,7 @@
 export * from "./types";
 
 export { createBridges } from "./bridge/index";
+export { createCantonOnboardingBridge, isCantonCurrencyBridge } from "./bridge/onboardingAdapter";
+export { isCantonAccount } from "./bridge/serialization";
 export type { CantonCoinConfig } from "./config";
 export { isAccountEmpty } from "./helpers";
-export { isCantonAccount } from "./bridge/serialization";
